code/requirements.txt
Flask==3.0.3
boto3==1.35.14
python-dotenv==1.0.1